split & cleanup BoDi.cs
