+++ Reqnroll/BoDi/IContainedInstance.cs
@@ -1,0 +1,6 @@
+﻿namespace Reqnroll.BoDi;
+
+public interface IContainedInstance
+{
+    IObjectContainer Container { get; }
+}
